Fetch the message body on click to avoid performance issues
Changed the logic of the MessageList.php component to
fetch only the message headers and defer content
fetching to the user click.

Created an html parser and web page scraper helper methods in
the Helpers trait.

Included bootstrap icons cdn in app.css

Fixed an issue with  the date parser in the HasMailboxConnection
trait.